Fixed some bugs and create Debug folder with Instrumentor which will be present soon
