+++ Pika/Pika.vcxproj
@@ -152,7 +152,6 @@
     </PostBuildEvent>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="src\Pika\Renderer\Renderer2D.h" />
     <ClInclude Include="src\Pika.h" />
     <ClInclude Include="src\Pika\Core\Application.h" />
     <ClInclude Include="src\Pika\Core\Core.h" />
@@ -183,6 +182,7 @@
     <ClInclude Include="src\Pika\Renderer\Camera2DController.h" />
     <ClInclude Include="src\Pika\Renderer\GraphicsContext.h" />
     <ClInclude Include="src\Pika\Renderer\RenderCommand.h" />
+    <ClInclude Include="src\Pika\Renderer\Renderer2D.h" />
     <ClInclude Include="src\Pika\Renderer\Renderer3D.h" />
     <ClInclude Include="src\Pika\Renderer\RendererAPI.h" />
     <ClInclude Include="src\Pika\Renderer\Shader.h" />
@@ -608,7 +608,6 @@
     <ClInclude Include="vendor\stb_image\stb_image.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="src\Pika\Renderer\Renderer2D.cpp" />
     <ClCompile Include="src\Pika\Core\Application.cpp" />
     <ClCompile Include="src\Pika\Core\Layer.cpp" />
     <ClCompile Include="src\Pika\Core\LayerStack.cpp" />
@@ -628,6 +627,7 @@
     <ClCompile Include="src\Pika\Renderer\Camera2DController.cpp" />
     <ClCompile Include="src\Pika\Renderer\GraphicsContext.cpp" />
     <ClCompile Include="src\Pika\Renderer\RenderCommand.cpp" />
+    <ClCompile Include="src\Pika\Renderer\Renderer2D.cpp" />
     <ClCompile Include="src\Pika\Renderer\Renderer3D.cpp" />
     <ClCompile Include="src\Pika\Renderer\RendererAPI.cpp" />
     <ClCompile Include="src\Pika\Renderer\Shader.cpp" />
