+++ Pika/src/Pika/Renderer/Renderer2D.h
@@ -13,7 +13,8 @@ namespace Pika
 		static void Init();
 		static void BeginScene(Camera2DController& vCameraController);
 		static void EndScene();
-		static void Submit(const Shader* vShader, const VertexArray* vData, const glm::mat4 vTransform = glm::mat4(1.0f)); //submit Scene info
+
+		//static
 	};
 
 }
