+++ Sandbox/imgui.ini
@@ -4,9 +4,8 @@ Size=400,400
 Collapsed=0
 
 [Window][Dear ImGui Demo]
-ViewportPos=-328,449
-ViewportId=0xE927CF2F
-Size=422,1061
+Pos=1498,19
+Size=422,936
 Collapsed=0
 
 [Window][Dear ImGui Debug Log]
@@ -21,7 +20,7 @@ Size=608,584
 Collapsed=0
 
 [Window][DockSpace Demo]
-Size=1920,1080
+Size=1920,1061
 Collapsed=0
 
 [Window][Dear ImGui ID Stack Tool]
@@ -54,12 +53,19 @@ Size=500,260
 Collapsed=0
 
 [Window][Color Board]
-Pos=160,90
-Size=329,70
+Pos=144,260
+Size=329,36
 Collapsed=0
 
+[Window][Color]
+Pos=529,157
+Size=329,70
+Collapsed=0
+DockId=0x00000001,0
+
 [Docking][Data]
-DockSpace     ID=0x3BC79352 Pos=294,336 Size=1920,1061 Split=Y Selected=0xE87781F4
+DockNode      ID=0x00000001 Pos=537,188 Size=329,70
+DockSpace     ID=0x3BC79352 Pos=8,50 Size=1920,1042 Split=Y Selected=0xE87781F4
   DockNode    ID=0x00000003 Parent=0x3BC79352 SizeRef=1920,903 Split=X
     DockNode  ID=0x00000005 Parent=0x00000003 SizeRef=694,903 Selected=0xCB2F3305
     DockNode  ID=0x00000006 Parent=0x00000003 SizeRef=698,903 CentralNode=1
